Refactors atom data import path organization
Moves parse_atom_data import from model module to dedicated atom_data module and removes unused legacy macro atom imports.

Improves code organization by consolidating atom data parsing functionality under a more appropriate module structure.
